+++ src/beastsofchaos/Chimera.cpp
@@ -129,7 +129,7 @@ namespace BeastsOfChaos {
         if (player == owningPlayer()) {
             // Fiery Breath
             if (m_shootingTarget) {
-                if (distanceTo(m_shootingTarget) <= (float) m_fieryBreath.range()) {
+                if (distanceTo(m_shootingTarget) <= (double) m_fieryBreath.range()) {
                     // Auto-hit and inflict mortal wounds.
                     Wounds breathDamage = {0, Dice::rollSpecial(g_damageTable[getDamageTableIndex()].m_fieryBreath)};
                     m_shootingTarget->applyDamage(breathDamage);
